// models/jsy-tymc-info.ts
export interface JsyTymcInfo {
  startStationId: string;
  endStationId: string;
  date: string;
  time: string;
  timeTables: MetroStationTimeTableTYMC["Timetables"];
  fareList: MetroODFareTYMC["Fares"];
}

// 含 jsy 客製欄位
export interface MetroStationTimeTableTYMC {
  RouteID: string;
  LineID: string;
  StationID: string;
  StationName: {
    Zh_tw: string; // 中文繁體名稱
    En: string; // 英文名稱
  };
  Direction: number; // 營運路線方向描述 : [0:'去程',1:'返程']
  DestinationStaionID: string;
  DestinationStationName: {
    Zh_tw: string;
    En: string;
  };
  Timetables: {
    Sequence: number; // 發車順序
    TrainNo: string; // 車次號碼(捷運通常沒有車次號碼)
    ArrivalTime: string; // 到站時刻(hh:mm)
    DepartureTime: string; // 發車時刻(hh:mm)
    jsyArrivalTime: string; // [客製] 抵達 endStation 時刻(hh:mm)
    jsyRunTime: string; // [客製] 行駛時間
    TrainType?: number; // 車種(0:不分車種, 1:普通車, 2:直達車)
    StoppingPatternID?: string; // 停站模式代碼
    jsyStoppingPatternName?: string; // [客製] 停站模式名稱
    jsyStoppingStationIdList?: string[]; // [客製] 停站車站清單
    jsyFareList: MetroODFareTYMC["Fares"]; // [客製] 票價
  }[];
  ServiceDay: {
    ServiceTag: string; // 營運日標籤
    Monday: boolean; // 星期一營運與否
    Tuesday: boolean;
    Wednesday: boolean;
    Thursday: boolean;
    Friday: boolean;
    Saturday: boolean;
    Sunday: boolean;
    NationalHolidays: boolean; // 國定假日營運與否
  };
  SpecialDays: {
    SaterDate: string; // 開始時間
    EndDate: string; // 結束時間
    Description: string; // 描述
  }[];
  SrcUpdateTime: string; // ISO8601格式
  UpdateTime: string; // ISO8601格式
  VersionID: number;
}
// ...
export interface MetroODFareTYMC {
  OriginStationID: string;
  OriginStationName: {
    Zh_tw: string;
    En: string;
  };
  DestinationStationID: string;
  DestinationStationName: {
    Zh_tw: string;
    En: string;
  };
  Fares: Array<{
    TicketType: number; // 1:一般票 2:來回票 3:電子票證 4:回數票 5:定期票30天 6:定期票60天
    FareClass: number; // 1:成人 2:學生 3:孩童 4:敬老 5:愛心 6:愛心孩童 7:愛心優待/愛心陪伴 8:團體
    SaleType: string; // 1:現場櫃台 2:現場機器 3:線上 99:其他
    CitizenCode: string; // TPE:臺北市 NWT:新北市 OTHERS:其他
    Price: number; // 新台幣
  }>;
  TravelTime: number; // 分鐘
  TravelDistance: number; // 公里
  SrcUpdateTime: string; // ISO8601: yyyy-MM-ddTHH:mm:sszzz
  UpdateTime: string; // ISO8601: yyyy-MM-ddTHH:mm:sszzz
  VersionID: number;
}